backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Email backend my portfolio website"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "fastapi-mail>=1.4.2",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
